backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-decouple>=3.8
